Refactor Lightstep code
